Rewrite Git structure guidance in English
Instead of a list of bullet points, explain in plain English how to
write a Git commit message, with a little more context as to the
reasoning behind it.

This duplicates some of the explanation from the guides we link to, such
as the explanation on why to wrap to 72 characters, but also covers
reasoning we hadn't documented before, like why you shouldn't end the
first line with a full stop.