Move guidance on external PRs from opsmanual
This is a wholesale copy of the page on the opsmanual about reviewing external PRs, which I will remove once this commit is on master. It makes sense for this information to be together with our other PR guidance, and it would be better for this to be public rather than on the internal-only opsmanual.

There is some overlap between what I am adding here and what is already in place on this page; however, I think it's important to add the information as-is as a straight move and we can then edit the rest of the document later.