Replace question with title
This heads up the section with what it's meant to cover. It also, instead of the
now unnecessary outline of what follows, adds a line to make it clear you can
ask for help. I'm still not sure this should be needed, but erring on the side
of welcoming new Git users.