This is a niche example
I am making these changes to this styleguide because it is too long and presents
an unbroken wall of bullet points which is too much to ask someone to read.

You don't need to suggest a way to review the PR in every case. It can be very
helpful to suggest removing whitespace, for example, but only in a small number
of cases. This guidance is not the place to cover every edge case.