Guidance for calling methods with multiple args
I've noticed various styles being used throughout the projects and
wanted to try and settle upon some recognised style for calling methods
with multiple arguments spanning multiple lines. My preferred style is
the one described as `#good`. I know others prefer `#bad`.

My view is the `#good` style apes our guidance for literal arrays and
hashes and adds readability. Often what can happen is those arguments
are indented at the opening paren level and can be missed or obscured
for those of us with 80 (and for some measures greater) char terminals.

If this gets general approval, I'll add a cop to `alphagov/govuk-lint`
and revisit this guidance with a referral to said cop.