Clarify wording about Smart Answers in use-of-READMEs.md
In the list of bad examples it wasn't obvious to me that the link to the Smart
Answers README was for an old version of the app. The previous wording implied
to me that the current version of the Smart Answer README is "bad" when it is
not.

Likewise in the list of good examples, it wasn't obvious to me what "modern"
Smart Answers meant.