Add licensing styleguide
We made lots of updates a few years ago to make sure everything had Crown Copyright, and most of our repos have licence files, but we don't consistently do this correctly; for example some repositories do not include the title of the licence so you have to read all the text to be sure what it is.

This commit adds a styleguide to clarify what the licence should look like.